requirements.txt
# CDK 
aws-cdk-lib==2.9.0
constructs>=10.0.0,<11.0.0
